% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{analyze_all_files}
\alias{analyze_all_files}
\title{Analyze all files}
\usage{
analyze_all_files(file_info)
}
\arguments{
\item{file_info}{File information from scan_directory}
}
\value{
A list with analysis results for all files
}
\description{
This function analyzes all files in the provided file information
}
