% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{analyze_csv}
\alias{analyze_csv}
\title{Analyze a CSV file}
\usage{
analyze_csv(file_info)
}
\arguments{
\item{file_info}{File information from scan_directory}
}
\value{
A list with analysis results
}
\description{
This function analyzes the structure and content of a CSV file
}
