% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{analyze_excel}
\alias{analyze_excel}
\title{Analyze an Excel file}
\usage{
analyze_excel(file_info)
}
\arguments{
\item{file_info}{File information from scan_directory}
}
\value{
A list with analysis results
}
\description{
This function analyzes the structure and content of an Excel file
}
