% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{detect_inconsistencies}
\alias{detect_inconsistencies}
\title{Detect inconsistencies across files}
\usage{
detect_inconsistencies(file_summaries)
}
\arguments{
\item{file_summaries}{Analysis results from analyze_all_files}
}
\value{
A list with inconsistency information
}
\description{
This function identifies inconsistencies in field types across files
}
