% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{enhance_harmonization_plan}
\alias{enhance_harmonization_plan}
\title{Enhance harmonization plan with key variable information}
\usage{
enhance_harmonization_plan(harmonization_plan, key_variables)
}
\arguments{
\item{harmonization_plan}{Harmonization plan from generate_harmonization_plan}

\item{key_variables}{List of identified key variables}
}
\value{
Enhanced harmonization plan
}
\description{
Adds information about potential key variables to the harmonization plan
}
