% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{find_common_fields}
\alias{find_common_fields}
\title{Find common fields across files}
\usage{
find_common_fields(file_summaries)
}
\arguments{
\item{file_summaries}{Analysis results from analyze_all_files}
}
\value{
A list with common field information
}
\description{
This function identifies fields that appear in multiple files
}
