% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{generate_harmonization_plan}
\alias{generate_harmonization_plan}
\title{Generate a harmonization plan}
\usage{
generate_harmonization_plan(report)
}
\arguments{
\item{report}{Inspection report from inspect_data}
}
\value{
A list with the harmonization plan
}
\description{
Creates a plan for harmonizing multiple dataframes based on inspection results
}
