% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{generate_report}
\alias{generate_report}
\title{Generate a comprehensive report}
\usage{
generate_report(file_info, file_summaries)
}
\arguments{
\item{file_info}{File information from scan_directory}

\item{file_summaries}{Analysis results from analyze_all_files}
}
\value{
A list with the comprehensive report
}
\description{
This function generates a comprehensive report of the data analysis
}
