% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{harmonize_dataframes}
\alias{harmonize_dataframes}
\title{Harmonize dataframes}
\usage{
harmonize_dataframes(dataframes, harmonization_plan, output_dir)
}
\arguments{
\item{dataframes}{List of dataframes to harmonize}

\item{harmonization_plan}{Plan generated by generate_harmonization_plan}

\item{output_dir}{Directory where to save harmonized files}
}
\value{
List of harmonized dataframes
}
\description{
Applies a harmonization plan to standardize dataframes
}
