% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\docType{package}
\name{harmonizeR}
\alias{odgharmonizR}
\alias{odgharmonizR-package}
\alias{harmonizeR}
\title{harmonizeR: Data Harmonization Tools for Swiss Open Data}
\description{
Tools for harmonizing, standardizing, and linking datasets from various 
sources, with special support for Swiss open data formats and geospatial services.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
