% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{identify_key_variables}
\alias{identify_key_variables}
\title{Identify potential key variables for linking datasets}
\usage{
identify_key_variables(dataframes, report)
}
\arguments{
\item{dataframes}{List of dataframes}

\item{report}{Inspection report}
}
\value{
List of potential key variables with their properties
}
\description{
Analyzes dataframes to identify variables that could be used for linking
}
