% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{link_datasets}
\alias{link_datasets}
\title{Link datasets}
\usage{
link_datasets(dataframes, key_variables, output_dir)
}
\arguments{
\item{dataframes}{List of harmonized dataframes}

\item{key_variables}{List of key variables from identify_key_variables}

\item{output_dir}{Directory where to save linked files}
}
\value{
A list containing:
  \item{linked_data}{The linked dataframe}
  \item{link_report}{A report on the linking process}
}
\description{
Links multiple dataframes based on key variables
}
