% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{load_files}
\alias{load_files}
\title{Load files based on inspection report}
\usage{
load_files(report, directory_path)
}
\arguments{
\item{report}{A list containing the inspection report data}

\item{directory_path}{Character string with path to the directory containing data files}
}
\value{
A list of dataframes
}
\description{
Loads data files according to the information in an inspection report
}
