% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{load_inspection_report}
\alias{load_inspection_report}
\title{Load inspection report from JSON file}
\usage{
load_inspection_report(report_path)
}
\arguments{
\item{report_path}{Character string with path to the JSON file}
}
\value{
A list with the loaded report
}
\description{
Load inspection report from JSON file
}
