% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{load_wfs}
\alias{load_wfs}
\title{Load data from a Web Feature Service (WFS)}
\usage{
load_wfs(wfs_url, feature_type)
}
\arguments{
\item{wfs_url}{Character string with the URL of the WFS service}

\item{feature_type}{Character string with the name of the feature type to load}
}
\value{
An sf object with the loaded features or NULL if loading fails
}
\description{
Connects to a WFS service and retrieves features for a specified feature type.
}
\examples{
\dontrun{
wfs_data <- load_wfs(
  "https://data.geo.admin.ch/api/stac/v0.9/collections/ch.swisstopo.swissboundaries3d-gemeinde-flaeche.fill/items",
  "ch.swisstopo.swissboundaries3d-gemeinde-flaeche.fill"
)
}
}
