% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{load_wms}
\alias{load_wms}
\title{Load data from a Web Map Service (WMS)}
\usage{
load_wms(wms_url, layer_name)
}
\arguments{
\item{wms_url}{Character string with the URL of the WMS service}

\item{layer_name}{Character string with the name of the layer to load}
}
\value{
A dataframe with layer metadata or NULL if loading fails
}
\description{
Connects to a WMS service and retrieves metadata for a specified layer.
}
\examples{
\dontrun{
wms_data <- load_wms(
  "https://wms.geo.admin.ch/",
  "ch.swisstopo.pixelkarte-farbe"
)
}
}
