% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{might_be_date}
\alias{might_be_date}
\title{Check if a column might contain dates}
\usage{
might_be_date(values)
}
\arguments{
\item{values}{Vector of values to check}
}
\value{
Logical indicating if the column might contain dates
}
\description{
This function samples values from a column and checks if they appear to be dates
}
\keyword{internal}
