% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{parse_wms_capabilities}
\alias{parse_wms_capabilities}
\title{Parse a WMS GetCapabilities response}
\usage{
parse_wms_capabilities(capabilities_xml)
}
\arguments{
\item{capabilities_xml}{Character string with the XML response from a WMS GetCapabilities request}
}
\value{
A list of layer information, each containing name, title, abstract, and bounding box
}
\description{
Extracts layer information from a WMS GetCapabilities XML response.
}
