% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeR-package.r
\name{save_report}
\alias{save_report}
\title{Save report to JSON file}
\usage{
save_report(report, file_path = "data_inspection_report.json")
}
\arguments{
\item{report}{Report from generate_report}

\item{file_path}{Path where to save the JSON file}
}
\value{
Invisibly returns the path to the saved file
}
\description{
This function saves the analysis report to a JSON file
}
